Serialization クラスを SaveData 以外のクラスからも使用するため外部に作成
