+++ KTest/Assets/Scripts/Common/Serialization.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 44795b08afacb494f83b92f0fd51b460
+timeCreated: 1503812581
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
